// front/src/gen/model/district.ts
/**
 * Sharing taxi api docs
 * Api description
 *
 * OpenAPI spec version: v1
 * Contact: contact@snippets.local
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Port } from './port';


export interface District { 
    readonly id?: number;
    ports: Array<Port>;
    name: string;
}