// front/src/gen/model/order.ts
/**
 * Sharing taxi api docs
 * Api description
 *
 * OpenAPI spec version: v1
 * Contact: contact@snippets.local
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Request } from './request';


export interface Order { 
    readonly id?: number;
    request?: Request;
    readonly created_datetime?: Date;
    date?: Date;
    readonly order_status?: Order.OrderStatusEnum;
    readonly user?: number;
    from_port: number;
    to_port: number;
}
export namespace Order {
    export type OrderStatusEnum = 'in_search' | 'waiting' | 'active' | 'finished' | 'cancelled' | 'route_error' | 'no_vehicles';
    export const OrderStatusEnum = {
        InSearch: 'in_search' as OrderStatusEnum,
        Waiting: 'waiting' as OrderStatusEnum,
        Active: 'active' as OrderStatusEnum,
        Finished: 'finished' as OrderStatusEnum,
        Cancelled: 'cancelled' as OrderStatusEnum,
        RouteError: 'route_error' as OrderStatusEnum,
        NoVehicles: 'no_vehicles' as OrderStatusEnum
    };
}